requirements.txt
numpy
pillow
tqdm
scikit-learn
matplotlib
torch
torchvision
facenet-pytorch
